+++ tests/strategies/test_deepspeed_strategy.py
@@ -558,6 +558,10 @@ class ModelParallelClassificationModel(LightningModule):
         if not hasattr(self, "model"):
             self.configure_sharded_model()
 
+        # Lightning saves the lr schedulers, but DeepSpeed saves the optimizer states separately
+        assert len(checkpoint["lr_schedulers"]) == 1
+        assert "optimizer_states" not in checkpoint
+
 
 class ManualModelParallelClassificationModel(ModelParallelClassificationModel):
     @property